server/pyproject.toml
[project]
name = "img-srch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.5"
dependencies = [
    "fastapi>=0.115.6",
    "numpy>=2.2.3",
    "pydantic>=2.10.4",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
    "torch>=2.7.0",
    "open-clip-torch>=2.32.0",
    "Pillow>=10.0.0",
]

[tool.uv.pip]
torch-backend = "auto"

[tool.uv.workspace]
members = ["qt"]
[tool.pyrefly]
project_includes = ["**/*"]
project_excludes = [
    "**/.[!/.]*",
    '**/*venv/**\*',
]
